Move internal struct definitionis to impl file
Make the public header even cleaner
